Baseline estimation (#38)
Baseline apps all pass under estimation. This also lead to all but 3 complete full synthesis. Next step is to evaluate these apps with golden testing which is prevented by #34.

Some issues which contain assumptions and errors found in this stage are, #35, #44, #26, #23 and #43. Trip count pragma estimates are not accurate which affect HLS estimated clock cycle details.