Centralize batch output
This is a pretty significant change to the way we organize all the
files: both the state we need to save between submission and collection,
and the actual results we download from the server. The idea is that
every batch we submit gets to own its own directory under `results/`.
That includes the list of jobs (and any failure information), and
eventually the files we download and such. This way, batches are a
little more independent/idempotent: you can submit multiple batches and
have them running in parallel, and collect their results when you need
to.

The `jobs.txt` file does not need to be clobbered when you submit a new
batch. Instead, we now have a `last_batch.txt` thing that makes it easy
to interact with the most recent batch you submitted, but it's also cool
to collect data from *any* batch you submitted in the past.