Start summarize script
This is a script to convert the information-rich `results.json` to a far
more readable `summary.csv` with just the information humans actually
want to know about benchmarks.

I'm not certain I have actually identified the right information to
include in this summary, but this is a first step. In particular, I
don't understand the differences between the "perf" and "synthesis"
outputs for the same statistics. Sometimes they're exactly the same;
sometimes they're quite different. Is one more trustworthy than the
other?