Change the names of the reports we have so far
Now the two things we get are called:
- `hls`, which comes from a file called `sds_foo.rpt` where `foo` is the
name of the hardware kernel function. This seems to be produced at "HLS
time," i.e., during the shorter process of compiling C++ to RTL. It
*seems* to be the same regardless of whether we're in estimation mode or
not.
- `est`, which comes from `perf.est`. This seems to be the *actual*
result of estimation mode, and it analyzes the "entire" design (instead
of just an individual kernel function). It is also less detailed.

Still missing: `synth` or something, which we'll get from the actual EDA
flow (RTL -> bitstream).